k8s-watcher/requirements.txt
kubernetes>=29.0.0
kafka-python>=2.0.2
urllib3>=2.0.0
